backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
langchain
langchain-openai
python-dotenv